IDE-2252 add "Add JSP Validation Support" to projects popup menu

+++ tools/plugins/com.liferay.ide.gradle.core/src/com/liferay/ide/gradle/core/ModuleCoreUtil.java
@@ -16,13 +16,19 @@
 package com.liferay.ide.gradle.core;
 
 import com.liferay.ide.core.util.FileUtil;
+import com.liferay.ide.project.core.util.ProjectUtil;
 
 import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IProjectDescription;
 import org.eclipse.core.resources.IResource;
+import org.eclipse.core.resources.IResourceVisitor;
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.OperationCanceledException;
 
@@ -32,7 +38,22 @@ import org.eclipse.core.runtime.OperationCanceledException;
 public class ModuleCoreUtil
 {
 
-    public static void addFacets( final IProject project, IProgressMonitor monitor ) throws CoreException
+    public static void addFacetsIfNeeded( final File projectLocation, IProgressMonitor monitor ) throws CoreException
+    {
+        final List<IProject> projects = new ArrayList<IProject>();
+
+        ProjectUtil.collectProjectsFromDirectory( projects, projectLocation );
+
+        for( final IProject project : projects )
+        {
+            if( hasJsp( project ) )
+            {
+                addFacets( project, monitor );
+            }
+        }
+    }
+
+    private static void addFacets( IProject project, IProgressMonitor monitor ) throws CoreException
     {
         addNature( project, "org.eclipse.wst.common.modulecore.ModuleCoreNature", monitor );
         addNature( project, "org.eclipse.wst.common.project.facet.core.nature", monitor );
@@ -54,7 +75,10 @@ public class ModuleCoreUtil
 
         File compoment = new File( project.getLocation().toFile(), ".settings/org.eclipse.wst.common.component" );
 
-        FileUtil.writeFile( compoment, finalContent.getBytes(), project.getName() );
+        if( !compoment.exists() )
+        {
+            FileUtil.writeFile( compoment, finalContent.getBytes(), project.getName() );
+        }
     }
 
     private static void addNature( IProject project, String natureId, IProgressMonitor monitor ) throws CoreException
@@ -85,6 +109,56 @@ public class ModuleCoreUtil
                 monitor.worked( 1 );
             }
         }
+    }
+
+    private static boolean hasJsp( IProject project )
+    {
+        final List<String> list = new ArrayList<String>();
+
+        try
+        {
+            final IPath projectPath = project.getLocation().append( ".project" );
+
+            final File projectFile = projectPath.toFile().getCanonicalFile();
+
+            project.accept( new IResourceVisitor()
+            {
+
+                @Override
+                public boolean visit( IResource resource ) throws CoreException
+                {
+                    try
+                    {
+                        IPath childProject = resource.getLocation().append( ".project" );
+
+                        File childFile = childProject.toFile().getCanonicalFile();
+
+                        //don't check child project
+                        if( childFile.exists() && !projectFile.equals( childFile ) )
+                        {
+                            return false;
+                        }
+
+                        String path = resource.getLocation().toPortableString();
+
+                        if( path.contains( "resources/META-INF/resources" ) && resource.getName().endsWith( ".jsp" ) )
+                        {
+                            list.add( path );
+                        }
+                    }
+                    catch( IOException e )
+                    {
+                    }
+
+                    return true;
+                }
+            } );
+        }
+        catch( Exception e )
+        {
+        }
+
+        return !list.isEmpty();
     }
 
     private static boolean hasNature( IProject project, String natureId )
